modify compare_states
