+++ proxy_scripts/ci_check.sh
@@ -42,8 +42,6 @@ elif [[ $M == "testnew" ]]; then
     cargo test --package proxy_tests --test proxy config
     cargo test --package proxy_tests --test proxy flashback
     cargo test --package proxy_tests --test proxy server_cluster_test
-    # tests based on new-mock-engine-store, for some tests not available for new proxy
-    cargo test --package proxy_tests --test proxy proxy
 elif [[ $M == "debug" ]]; then
     # export RUSTC_WRAPPER=~/.cargo/bin/sccache
     export ENGINE_LABEL_VALUE=tiflash
